refactor: relocate and refactor test files
- move test files from `tests/tests/` to `benchmark/test/`
- move vitest config from `tests/` to `benchmark/`
- update import paths to match new directory structure
- fix optional chaining in openai-embedding test
- remove `tests.ts` command file (previous commit)
